+++ lib_feature/src/features/brightness_distribution.rs
@@ -7,7 +7,7 @@ use crate::features::{Feature, MinimumValue};
 /// In X direction, find for a specified length of timeframes the area (left 0, center 1, right 2)
 /// where it is brightest for each time frame, i.e. where the difference to the minimum is the minimal (Minimum - current).
 /// Its called float because it can also be also take values between the areas.
-pub fn calc_brightness_distribution_float_x(length: usize, gesture: &Gesture) -> Vec<f64> {
+pub fn calc_brightness_distribution_float_x(length: usize, gesture: &Gesture) -> Vec<f32> {
     assert!(gesture.frames.len() >= length);
     let mut result = Vec::new();
     let min_frame = MinimumValue::calculate(gesture);
@@ -20,15 +20,15 @@ pub fn calc_brightness_distribution_float_x(length: usize, gesture: &Gesture) ->
     }
 
     // Tries to shorten to result to the specified length
-    let merge_threshold = result.len() as f64 / (length as f64);
+    let merge_threshold = result.len() as f32 / (length as f32);
     let mut values = Vec::new();
     let mut perma_result = Vec::with_capacity(length);
     for i in 0..result.len() {
         values.push(result[i]);
-        if ((i + 1) as f64) < merge_threshold * ((perma_result.len() + 1) as f64) {
+        if ((i + 1) as f32) < merge_threshold * ((perma_result.len() + 1) as f32) {
             continue;
         }
-        perma_result.push((values.iter().sum::<usize>() as f64) / (values.len() as f64));
+        perma_result.push((values.iter().sum::<usize>() as f32) / (values.len() as f32));
         values.clear();
     }
     perma_result
@@ -37,7 +37,7 @@ pub fn calc_brightness_distribution_float_x(length: usize, gesture: &Gesture) ->
 /// In Y direction, find for a specified length of timeframes the area (top 0, center 1, bottom 2)
 /// where it is brightest for each time frame, i.e. where the difference to the minimum is the minimal (Minimum - current).
 /// Its called float because it can also be also take values between the areas.
-pub fn calc_brightness_distribution_float_y(length: usize, gesture: &Gesture) -> Vec<f64> {
+pub fn calc_brightness_distribution_float_y(length: usize, gesture: &Gesture) -> Vec<f32> {
     assert!(gesture.frames.len() >= length);
     let mut result = Vec::new();
     let min_frame = MinimumValue::calculate(gesture);
@@ -48,15 +48,15 @@ pub fn calc_brightness_distribution_float_y(length: usize, gesture: &Gesture) ->
         result.push(vec![row1, row2, row3].into_iter().enumerate()
             .min_by(|&(_, left), &(_, right)| left.cmp(&right)).unwrap().0);
     }
-    let merge_threshold = result.len() as f64 / (length as f64);
+    let merge_threshold = result.len() as f32 / (length as f32);
     let mut values = Vec::new();
     let mut perma_result = Vec::with_capacity(length);
     for i in 0..result.len() {
         values.push(result[i]);
-        if ((i + 1) as f64) < merge_threshold * ((perma_result.len() + 1) as f64) {
+        if ((i + 1) as f32) < merge_threshold * ((perma_result.len() + 1) as f32) {
             continue;
         }
-        perma_result.push((values.iter().sum::<usize>() as f64) / (values.len() as f64));
+        perma_result.push((values.iter().sum::<usize>() as f32) / (values.len() as f32));
         values.clear();
     }
     perma_result
@@ -75,12 +75,12 @@ pub fn calc_brightness_distribution_float_xy(length: usize, gesture: &Gesture) -
     }
 
     // Tries to shorten to result to the specified length
-    let merge_threshold = result.len() as f64 / (length as f64);
+    let merge_threshold = result.len() as f32 / (length as f32);
     let mut values = Vec::new();
     let mut perma_result = Vec::with_capacity(length);
     for i in 0..result.len() {
         values.push(result[i]);
-        if ((i + 1) as f64) < merge_threshold * ((perma_result.len() + 1) as f64) {
+        if ((i + 1) as f32) < merge_threshold * ((perma_result.len() + 1) as f32) {
             continue;
         }
         perma_result.push(values.iter().min().cloned().unwrap());
@@ -115,12 +115,12 @@ pub fn calc_brightness_distribution_float_xy_geom(length: usize, gesture: &Gestu
     }
 
     // Tries to shorten to result to the specified length
-    let merge_threshold = result.len() as f64 / (length as f64);
+    let merge_threshold = result.len() as f32 / (length as f32);
     let mut values = Vec::new();
     let mut perma_result = Vec::with_capacity(length);
     for i in 0..result.len() {
         values.push(result[i]);
-        if ((i + 1) as f64) < merge_threshold * ((perma_result.len() + 1) as f64) {
+        if ((i + 1) as f32) < merge_threshold * ((perma_result.len() + 1) as f32) {
             continue;
         }
         let (x, y) = values.iter().fold((0, 0), |(acc_x, acc_y), (x, y)| (acc_x + x, acc_y + y));
@@ -166,12 +166,12 @@ pub fn calc_brightness_distribution_float_xy_quadrant(length: usize, gesture: &G
     }
 
     // Tries to shorten to result to the specified length
-    let merge_threshold = result.len() as f64 / (length as f64);
+    let merge_threshold = result.len() as f32 / (length as f32);
     let mut values = Vec::new();
     let mut perma_result = Vec::with_capacity(length);
     for i in 0..result.len() {
         values.push(result[i]);
-        if ((i + 1) as f64) < merge_threshold * ((perma_result.len() + 1) as f64) {
+        if ((i + 1) as f32) < merge_threshold * ((perma_result.len() + 1) as f32) {
             continue;
         }
         let (x, y) = values.iter().fold((0, 0), |(acc_x, acc_y), (x, y)| (acc_x + x, acc_y + y));
